Add files via upload
As the normal graph (ie BFS and DFS) has the time complexity of O(V+E), where v:vertex and e:edges. Here we are having )(V)^V, which is worse compare to the previous one. But to obtain the solution we are looking for is the solution. lets look close into it......